Series of changes
Enable logs for 1 year
Runstats need to run after everyday after backup
Update the timestamp of audit file
Uniq db names from the database directory
Add ICD api key

+++ instance-applications/120-ibm-db2u-database/files/runstats_rebind.sh
@@ -27,11 +27,13 @@ if [ -f /mnt/backup/bin/.PROPS ]
 then
     . /mnt/backup/bin/.PROPS
     DOW=$(date | awk '{print $1}')
+    : '
     if [ ${DOW} != ${DAYOFFULL} ]
     then
         echo "Runstats runs only on Day of Full backup. . . !!! Exiting !"
         exit 0
     fi
+    '
 fi
 
 # -- Standard Parameters
@@ -64,7 +66,7 @@ fi
 RUNSTATS_TMP_FILE="${INSTANCE_HOME}/bin/.runstats.sql"
 REBIND_TMP_FILE="${INSTANCE_HOME}/bin/.rebind.sql"
 
-for DB in $(db2 list db directory | grep -B 5 Indirect | grep "Database name" | cut -d= -f2)
+for DB in $(db2 list db directory | grep -B5 "Indirect" | grep "Database name" |  awk '{ print $4 }' | sort -u  )
 do
     RUNSTATS_REBIND_LOG="${INSTANCE_HOME}/maintenance/logs/runstats_rebind_${DB}_${DATESTAMP}.log"
 
@@ -74,7 +76,7 @@ do
             rm ${RUNSTATS_TMP_FILE}
         fi
         if [[ -f ${REBIND_TMP_FILE} ]]; then
-            rm ${REBIND_TMP_FILE}
+            rm ${REBIND_TMP_FILE}   
         fi
 
         db2 connect to ${DB} | tee ${RUNSTATS_REBIND_LOG}
@@ -90,18 +92,22 @@ do
             echo -e "\nEnd processing of runstats @ ${DATESTAMP}" | tee -a ${RUNSTATS_REBIND_LOG}
             rm ${RUNSTATS_TMP_FILE}
 
-            # -- Rebind of the packages 
-            db2 -x "select 'rebind package \"' ||rtrim(PKGSCHEMA)||'\".\"'|| PKGNAME ||'\";' from syscat.packages 
-            where PKGSCHEMA not like 'SYSIBM%' and PKGSCHEMA not like 'NULL%' " > ${REBIND_TMP_FILE}
-            
-            # -- Execution of Rebind 
-            echo -e "\n ----------------------------------------------- " | tee -a ${RUNSTATS_REBIND_LOG}
-            echo -e "Begin processing of rebind @ ${DATESTAMP} ...\n" | tee -a ${RUNSTATS_REBIND_LOG}
-            db2 -txvf ${REBIND_TMP_FILE} | tee -a ${RUNSTATS_REBIND_LOG}
-            echo -e "\nEnd processing of rebind @ ${DATESTAMP}" | tee -a ${RUNSTATS_REBIND_LOG}
+            if [[ "${DOW}" == "${DAYOFFULL}" ]]; then
+
+                # -- Rebind of the packages 
+                db2 -x "select 'rebind package \"' ||rtrim(PKGSCHEMA)||'\".\"'|| PKGNAME ||'\";' from syscat.packages 
+                where PKGSCHEMA not like 'SYSIBM%' and PKGSCHEMA not like 'NULL%' " > ${REBIND_TMP_FILE}
                 
-            rm ${REBIND_TMP_FILE}
+                # -- Execution of Rebind 
+                echo -e "\n ----------------------------------------------- " | tee -a ${RUNSTATS_REBIND_LOG}
+                echo -e "Begin processing of rebind @ ${DATESTAMP} ...\n" | tee -a ${RUNSTATS_REBIND_LOG}
+                db2 -txvf ${REBIND_TMP_FILE} | tee -a ${RUNSTATS_REBIND_LOG}
+                echo -e "\nEnd processing of rebind @ ${DATESTAMP}" | tee -a ${RUNSTATS_REBIND_LOG}
+                    
+                rm ${REBIND_TMP_FILE}
+            fi
             db2 terminate
+
         fi
     fi
 done
